Add overload of RuntimeBundle::create for a group of IRFunctions
Summary:
This commit adds an overload of RuntimeBundle::create that allocates the
constants, placeholders and activations used by a group of IRFunctions in the
same one RuntimeBundle. It also slightly modifies IRGen for tensorviews to make
sure that there are no name collisions between tensorview symbols from
different IRFunctions in the common RuntimeBundle.

Reviewed By: nickgg

Differential Revision: D16870208

fbshipit-source-id: c9a4f5c61e1aaf2cbee865e5db51524f1c4acec4