[Backend] Add a default implementation of createDeviceManager
Description:
This commit adds a default implementation of
Backend::createDeviceManager which creates and returns a
DummyDeviceManager. Now, developers working on new backends do
not have to explicitly override it just to return a
DummyDeviceManager. This is useful for backends under development
that do not have DeviceManager implementations yet. Currently, a
DummyDeviceManager is created only for unregistered backends.

Testing:
Commented out override of Backend::createDeviceManager in the Intepreter
backend and ran ninja check. Tests can run and logs show the creation of
the DummyDeviceManager, but some tests fail because of some code that is
in the InterpreterDeviceManager but not in the DummyDeviceManager.