Enable saturateHost in image-classifier and address a couple resulting bugs (#3443)
Summary:
This PR enables saturate host in image-classifier, so now if multiple devices are created they will be used.
As a result of this change bugs in OpenCL and CPUFunctions were exposed this PR fixes those too.

Documentation:
Pull Request resolved: https://github.com/pytorch/glow/pull/3443

Test Plan: run image-classifier with num-devices >0 and backends= CPU, Interpreter, and OpenCL.

Differential Revision: D16941867

Pulled By: gcatron

fbshipit-source-id: 8fe1fb4d777c8dc5fa1b8d88e14fbee01dc0ed89