Fuse reshape and its inputs to glow (#3438)
Summary:
Import `aten::reshape` as well as `prim::ListConstruct` for int[] and `aten::size` in order to get that last couple of ops in resnet lowered.

Also this enable lowering of transpose an cleans up fusers.

This PR also introduces a bit of a problem. Before all outputs of Glow could be correctly mapped to PT tensors (even if not all inputs could). However now that we have `prim::ListConstruct` and `aten::size`, the glow subgraph can easily produce output IValues that are not tensors. We can handle this issue in a followup. We may need to add some metadata to each input/output placeholder to PyTorchModelLoader that explains how to map it to PyTorch values.

Documentation:
doxygen
Pull Request resolved: https://github.com/pytorch/glow/pull/3438

Test Plan:
CI
run resnet_example.py

Differential Revision: D16919605

fbshipit-source-id: a680e6f327dfb890753089e9fb1fe20601c20fbc