[NodeGen] Add the ability to use external user-defined types
Definitions of such user-defined types do not need to be known when compiling the NodeGen.cpp, but it should be known when compiling the generated files.  The types can be referenced in NodeGen by means of the MEMBER_TYPE_INFO macros as shown below. There is also MEMBER_TYPE_INFO_WITH_CMP for providing a custom comparator function, e.g. for pointer types.

The intended use in NodeGen is something like this:
```c++
  // Include the header providing the definitions of user-defined types into the generated header file.
  BB.includeHeader("glow/UserTypes/UserDefinedTypes.h");

  BB.newNode("CustomNode")
      .addMember(MEMBER_TYPE_INFO(UserDefinedExternalType), "F1")
      .addMember(MEMBER_TYPE_INFO_WITH_CMP(UserDefinedExternalType *, "isEqual"), "F2")
      .addMember(MEMBER_TYPE_INFO(UserDefinedExternalType &), "F3");
```