Added new priority_queue for ordering queded requests (#3416)
Summary:
This PR adds an optional priority field to runNetwork on HostManager. Also it adds a priority queue.
When a request is made of the HostManager the following happens:
The request is added to the priorityQueue if the max queue size limit has been reached return an error immediately.
If there are less active requests than the activeRequestLimit dispatchNextRun() is called.
If the activeLimit has been reached return.

Documentation: Updated in code documentation for new data structure and priority parameter in runNetwork.
Pull Request resolved: https://github.com/pytorch/glow/pull/3416

Test Plan: Added a new QueueTest to HostManagerTest.

Differential Revision: D16874009

Pulled By: gcatron

fbshipit-source-id: 643eddc1d0de229f78c973abd170a2f38a5e45d5