add FusedActivation field to Convolutions and remove HabanaConvolutionNode
Summary:
It's a common pattern among backends to fuse the activation (e.g. Relu) into Convolutions, which we have  handled previously with a Backend specific node (see HabanaConvolutionNode). Since we'll use this functionality in multiple backends, bringing the fused activation into ConvolutionNode - which allows us to remove HabanaConvolutionNode.

The FusedActivation field is currently only set during the transformPostLowering stage of the Habana backend. We assert if it's set on the other backends.

Reviewed By: opti-mix

Differential Revision: D16924415

fbshipit-source-id: 452e36c2cd03da0a6efb0c2e97ce21b1c45a9e40