Add the ability to use external user-defined types (#3448)
Summary:
add the ability to use external user-defined types in NodeGen.

This is quick proof-of-concept based on discussions with SplitInfinity and nickgg about adding fields of custom types (including enums) to Nodes and Instructions.

Documentation:
Definitions of such user-defined types do not need to be known when compiling the NodeGen.cpp, but it should be known when compiling the generated files.  The types can be referenced in NodeGen by means of the MEMBER_TYPE_INFO macros as shown below. There is also MEMBER_TYPE_INFO_WITH_CMP for providing a custom comparator function, e.g. for pointer types.

The intended use in NodeGen is something like this:
```c++
  // Include the header providing the definitions of user-defined types into the generated header file.
  BB.includeHeader("glow/UserTypes/UserDefinedTypes.h");
  ...
  // Define a node that uses user-defined types.
  BB.newNode("CustomNode")
      .addMember(MEMBER_TYPE_INFO(UserDefinedExternalType), "F1")
      .addMember(MEMBER_TYPE_INFO_WITH_CMP(UserDefinedExternalType *, "isEqual"), "F2")
      .addMember(MEMBER_TYPE_INFO(UserDefinedExternalType &), "F3");
```
Each user-defined type used this way should satisfy the following requirements:
   * There should be a hash function with a signature like `llvm::hash_code hash_value(const T)` which takes T by value, by reference or as a pointer, depending on the intended use.
  * There should be a stream output operator with a signature like `llvm::raw_ostream &operator<<(llvm::raw_ostream &os, const T);`, which takes T by value, by reference or as a pointer, depending on the intended use.
  * There should be a comparison operator `bool operator==(const T LHS, const T RHS)` (or a custom comparator function mentioned in MemberTypeInfo::cmpFn), which takes T by reference or by value depending on the intended use.

It is a responsibility of developer to define these functions for their custom types .
Pull Request resolved: https://github.com/pytorch/glow/pull/3448

Test Plan: ninja test

Reviewed By: nickgg

Differential Revision: D17071673

Pulled By: opti-mix

fbshipit-source-id: 27b688579270f8442d623b8693e25ea7584e9206