Add FP16 converter helper, and then support UInt8FusedQTy to UInt8FusedFP16QTy conversion with it. (#3420)
Summary:
We want to support converting UInt8FusedQTy to UInt8FusedFP16QTy when we do FP16 conversion. This does not fit naturally into the TypeAToTypeBFunctionConverter. Instead of making major modifications/redesigning it, I added a helper to wrap the TypeAToTypeBFunctionConverter with Float -> Float16 with a secondary pass to convert UInt8FusedQTy to UInt8FusedFP16QTy.

I additionally added support for the graph optimization for ConvertTo to merge a Constant UInt8FusedQTy followed by a ConvertTo UInt8FusedFP16QTy into the Constant.
Pull Request resolved: https://github.com/pytorch/glow/pull/3420

Test Plan: Added a test for the new graph optimization case that is supported. Added an OperatorTest and a conversion test to test that converting a FRWQ-SLWS to FP16 works correctly.

Reviewed By: rdzhabarov

Differential Revision: D16844779

Pulled By: jfix71

fbshipit-source-id: c71d5c903a24aa6a3bba6c3546564ab26bc2bcce