Clip before converting to fp16
Summary:
There's a hypothesis that FP16-converted nets need explicit clipping
of INFs to generate proper values.  To help test this hypothesis I've added a
precision config to clip FP16.

Note this is going to be pretty inefficient since it'll do min/max before and
after every FP16-converted node.  We can probably optimize a bunch of this
stuff away eventually but I suspect it's not worth doing just yet.

Reviewed By: tracelogfb

Differential Revision: D16723546

fbshipit-source-id: d0ecd9d5b23054fb85c8068642bc006df6081e52