[Graph] Only tile RHS of a BatchMatMul if LHS batch size is > 1
Summary:
This commit modifies Function::createBatchMatMul to insert a Tile
operation to make the first dimension (the batch dimension) of the RHS
operand equal to that of the LHS operand only if the latter is greater
than 1. If it is equal to 1, the Tile is not needed.

Test Plan:
This commit adds a new unit test for BatchMatMul in which the batch
dimension of the LHS operand is 1. All unit tests pass.