Only tile RHS of a BatchMatMul if LHS batch size is > 1 (#3520)
Summary:
**Summary**
This commit modifies `Function::createBatchMatMul` to insert a `Tile`
operation to make the first dimension (the batch dimension) of the RHS
operand equal to that of the LHS operand only if the latter is greater
than 1. If it is equal to 1, the `Tile` is not needed.
Pull Request resolved: https://github.com/pytorch/glow/pull/3520

Test Plan:
This commit adds a new unit test for `BatchMatMul` in which the batch
dimension of the LHS operand is 1. All unit tests pass.

Reviewed By: arunm-git, putivsky

Differential Revision: D17427721

Pulled By: SplitInfinity

fbshipit-source-id: 26e4166325f804ac6d096323d2bd7b855be10389