[Grad] Add support for differentiating BatchMatMul
Summary:
This commit adds support for differentiating BatchMatMul nodes.
Recall that the gradients with respect to both inputs of a MatMul node are computed
using MatMul nodes with the gradient with respect to the output of the
original MatMul and transposed versions of either input. Similarly, the
gradients with respect to both inputs of BatchMatMul node are computed
using BatchMatMul nodes with the (batched) gradient with respect to the
output of the original BatchedMatMul and transposed versions of either
input.

Testing:
This commit adds tests to GradCheckTest and GraphGradTest for
BatchMatMulNode. All tests pass.