[Graph] Make Module::addConstant private and delete Module::addPlaceholder
Summary:
Storage nodes (i.e. Constants and Placeholders) are supposed to be owned
by a Module. Accordingly, the preferred methods of creation for these
objects are Module::createConstant and Module::createPlaceholder.
Exposing public addConstant and addPlaceholder methods makes it possible
to create a Storage node and add it to multiple Modules. These methods
are not used frequently, so this commit makes one of them them private
and deletes the other in order to preclude this issue.

Testing:
This commit adds a new unit test to GraphTest to make sure that
Module::verify() can catch the case in which a Storage node is owned by
one Module but used by a Function in another Module.