[GraphOptz] Add helper function to check numerical correctness of optimizations
Summary:
This commit adds a helper function to GraphOptzTest that helps check
whether a Function and its optimized counterpart produce the same
results when given the same inputs. At the moment, the tests in
GraphOptzTest make sure that a Function is mutated as expected, but
don't check that the optimization is correct.

Testing:
This commit modifies one of the test cases in GraphOptzTest,
cancelTwoTransposes to use this new feature. All tests pass.