Make Module::addConstant private and delete Module::addPlaceholder (#3495)
Summary:
**Summary**
`Storage` nodes (i.e. `Constants` and `Placeholders`) are supposed to be owned
by a `Module`. Accordingly, the preferred methods of creation for these
objects are `Module::createConstant` and `Module::createPlaceholder`.
Exposing public `addConstant` and `addPlaceholder` methods makes it possible
to create a `Storage` node and add it to multiple `Modules`. These methods
are not used frequently, so this commit makes one of them them private
and deletes the other in order to preclude this issue.

**Testing**
This commit adds a new unit test to `GraphTest` to make sure that
`Module::verify()` can catch the case in which a `Storage` node is owned by
one `Module` but used by a `Function` in another `Module`.

**Fixes**
This PR fixes https://github.com/pytorch/glow/issues/3428.
Pull Request resolved: https://github.com/pytorch/glow/pull/3495

Differential Revision: D17510607

Pulled By: SplitInfinity

fbshipit-source-id: 74c42b97ba4c5097b0ae1af65bd959ed44a4be10