Not to use pip to install cmake in PYTORCH and CHECK_FORMAT branch. (#3545)
Summary:
Currently CI is down because of we are using pip install cmake which may cause some problem.
In this diff we will try to solve this problem by using apt-get or conda.

Since PYTORCH branch uses an older version of llvm, pip install cmake does not work for that. Therefore switch to use apt-get for that branch.
Pull Request resolved: https://github.com/pytorch/glow/pull/3545

Differential Revision: D17555624

Pulled By: zrphercule

fbshipit-source-id: 62085a024e8fb11af2d349a74489ff508e664a17