Add helper function to check numerical correctness of optimizations (#3535)
Summary:
**Summary**
This commit adds a helper function to `GraphOptzTest` that helps check
whether a `Function` and its optimized counterpart produce the same
results when given the same inputs. At the moment, the tests in
`GraphOptzTest` make sure that a `Function` is mutated as expected, but
don't check that the optimization is correct.

**Testing**
This commit modifies one of the test cases in `GraphOptzTest`,
`cancelTwoTransposes` to use this new feature. All tests pass.

**Fixes**
This PR fixes https://github.com/pytorch/glow/issues/3523.
Pull Request resolved: https://github.com/pytorch/glow/pull/3535

Differential Revision: D17553453

Pulled By: SplitInfinity

fbshipit-source-id: 7041cbad34c240f28fda51a7a0edc9e007bd6def