Add optimization to eliminate no-op TileNode (#3547)
Summary:
**Summary**
This commit adds a new graph optimization that eliminates `TileNodes` that
tile their input only once because this operation is a no-op.

**Testing**
This commit adds a unit test to `GraphOptzTest` that checks that a graph
containing a `TileNode` with count=1 is eliminated.

**Fixes**
This PR fixes https://github.com/pytorch/glow/issues/3522.
Pull Request resolved: https://github.com/pytorch/glow/pull/3547

Differential Revision: D17563282

Pulled By: SplitInfinity

fbshipit-source-id: 34fcf8f349cab211e2b7c8f50e60bec3941c7ffb