Add support for differentiating TileNode (#3527)
Summary:
**Summary**
This commit adds support for differentiating `TileNodes`. The gradient
with respect to the input of the `TileNode` is the sum of the slices of
the gradient with respect to the output of the `TileNode` that correspond
to the slices that were repeated and concatenated during the forward
pass of `TileNode`. This operation is performed by using a `ReshapeNode` to
modify the dimensions of the gradient with respect to the `TileNode` output
such that the tiling axis is replaced with `{numTiles, tileDim}` and then
adding all of the slices with a `BatchedReduceAdd` node on the axis
corresponding to `numTiles`.

**Testing**
This commit adds unit tests to `GraphGradTest` and `GradCheckTest` for
`TileNode`. The latter checks that the gradient is correct when tiling on
a nontrivial axis (i.e. not the first). All unit tests pass.
Pull Request resolved: https://github.com/pytorch/glow/pull/3527

Differential Revision: D17580514

Pulled By: SplitInfinity

fbshipit-source-id: 7ea8c76b6466394b7e35420dc6305aa51f960c63