Simplify OptimizeConversions
Summary:
OptimizeConversions was overly complex since it was trying to do its
own DCE, and not really taking advantage of idiomatic pattern matching or node
list traversal.  As such it wasn't a great example for others to follow.

Simplifying the code not only makes the logic clearer, but makes it trivial to
take care of the TODO item (optimizing non-narrowing cast chains with unequal
src/dest types, as in i8->i32->i16).

(Note: this ignores all push blocking failures!)

Reviewed By: jfix71

Differential Revision: D17595085

fbshipit-source-id: 3ed52f5f6eec10812524daa31625b1d0d1553250