Add support for differentiating BatchMatMul (#3529)
Summary:
**Summary**
This commit adds support for differentiating `BatchMatMul` nodes.
Recall that the gradients with respect to both inputs of a `MatMul` node are computed
using `MatMul` nodes with the gradient with respect to the output of the
original `MatMul` and transposed versions of either input. Similarly, the
gradients with respect to both inputs of `BatchMatMul` node are computed
using `BatchMatMul` nodes with the (batched) gradient with respect to the
output of the original `BatchedMatMul` and transposed versions of either
input.

**Testing**
This commit adds tests to `GradCheckTest` and `GraphGradTest` for
`BatchMatMulNode`. All tests pass.

**Dependencies**
This PR is stacked on top of https://github.com/pytorch/glow/issues/3527.
Pull Request resolved: https://github.com/pytorch/glow/pull/3529

Differential Revision: D17638410

Pulled By: SplitInfinity

fbshipit-source-id: 0142119cdeaff9a5e99c4d76911d6c348d5d1474