[GradCheckTest] Make pool tests test layout conversion
Summary:
This commit modifies the max and average pool tests in GradCheckTest to
test layout conversion as well. At present, NHWC -> NCHW layout
conversion for these test cases ends up becoming a Reshape, which can be
a no-op on the device.

Testing:
These tests are enabled for the Interpreter. All unit tests pass.