[GradCheckTest] Remove slice from average and max pool tests
Summary:
This commit removes the SliceNode from the gradient check tests for
average and max pooling because it is not needed in order to make the
test work. It was initially added to preserve the input type of the
FullyConnected layer, but that layer flattens whatever is given to it.

Testing:
All unit tests pass.