Remove slice from average and max pool tests (#3587)
Summary:
**Summary**
This commit removes the `SliceNode` from the gradient check tests for
average and max pooling because it is not needed in order to make the
test work. It was initially added to preserve the input type of the
`FullyConnected` layer, but that layer flattens whatever is given to it.

**Testing**
All unit tests pass.
Pull Request resolved: https://github.com/pytorch/glow/pull/3587

Differential Revision: D17727071

Pulled By: SplitInfinity

fbshipit-source-id: e335afd767516471190290b7385c1950b4f1090a