[OpenCL] Add src operand to maxpoolgrad kernel
Summary:
This commit adds a src operand to the OpenCL maxpoolgrad kernel. This
operand was added to the MaxPoolWithArgmaxGradInst class in an earlier
commit, but was not added to the corresponding OpenCL kernel. This
breaks the code in the OpenCL backend that sets kernel arguments using
operands from the corresponding instruction because it sets them in the order
of that they are declared in the instruction class.

Testing:
All unit tests pass.