Add src operand to maxpoolgrad kernel (#3599)
Summary:
**Summary**
This commit adds a src operand to the OpenCL maxpoolgrad kernel. This
operand was added to the `MaxPoolWithArgmaxGradInst` class in an earlier
commit, but was not added to the corresponding OpenCL kernel. This
broke the code in the OpenCL backend that sets kernel arguments using
operands from the corresponding instruction because it sets them in the order
of that they are declared in the instruction class.

**Testing**
All unit tests pass.
Pull Request resolved: https://github.com/pytorch/glow/pull/3599

Differential Revision: D17772008

Pulled By: SplitInfinity

fbshipit-source-id: 80a5c5819ff343c6a72e82b9cb4de87ca387354d