fix image classifier when tracing batches (#3596)
Summary:
The image-classifier has multiple paths to running the inference, and TraceEvents only worked in one of them (single image no batch). This fixes the other path to supply and merge a TraceContext so we can trace many runs.

![image](https://user-images.githubusercontent.com/701287/66227258-adadb680-e691-11e9-85ee-1497e92fb163.png)

This is a bit hacky, we should not have multiple paths in image-classifier but that's a bigger task (i'm looking at it). Would like to land this to unblock in the short term.

Documentation: NFC
Pull Request resolved: https://github.com/pytorch/glow/pull/3596

Test Plan: unit tests && tests/images/run.sh

Differential Revision: D17768309

Pulled By: nickgg

fbshipit-source-id: 6c391611637ef2b6e0436cacb78a13ef03446927