[GraphOptimizer] Sink Transpose below Arithmetic nodes when one operand is a Constant
Summary:
This commit extends the optimization that sinks Transpose nodes below Add
nodes to handle the case in which one of the operands of the Add node is
a Constant. This can arise if the backend chooses to replace Splat nodes
with Constants to avoid computing them at runtime.

Testing:
This commit adds a unit test that tests that a Transpose node is sunk below
an Add node when one operand is a Constant and the other is a Transpose node.