Sink Transpose below Arithmetic nodes when one operand is a Constant (#3621)
Summary:
**Summary**
This commit extends the optimization that sinks `Transpose` nodes below `Add`
nodes to handle the case in which one of the operands of the `Add` node is
a `Constant`.

**Testing**
This commit adds a unit test that tests that a `Transpose` node is sunk below
an `Add` node when one operand is a `Constant` and the other is a `Transpose` node.

```
$ ./tests/GraphOptzTest --gtest_filter=*sinkTransposeBelowArithmeticNodesWithConstantOperand*
Note: Google Test filter = *sinkTransposeBelowArithmeticNodesWithConstantOperand*
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from GraphOptz
[ RUN      ] GraphOptz.sinkTransposeBelowArithmeticNodesWithConstantOperand
[       OK ] GraphOptz.sinkTransposeBelowArithmeticNodesWithConstantOperand (5 ms)
[----------] 1 test from GraphOptz (5 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test suite ran. (5 ms total)
[  PASSED  ] 1 test.
```
Pull Request resolved: https://github.com/pytorch/glow/pull/3621

Differential Revision: D17935317

Pulled By: SplitInfinity

fbshipit-source-id: 913a47020b3312cd031ecfcf5e7748bfcca01ecf