Refactor buffer copies into DeviceManager (#3644)
Summary:
As a preparation for supporting device resident Tensors, move Tensor copies out of the OCLCompiledFunction and into the DeviceManager. Mostly this is just a move of the relevant code into the new class, but I had to move the KernelLaunches to the OCLDeviceBindings.

Fixed a few small TraceEventTest issues as well.

Documentation: NFC
Pull Request resolved: https://github.com/pytorch/glow/pull/3644

Test Plan: ran tests, image-classifier, and resnet-runtime with OCL

Differential Revision: D18018548

Pulled By: nickgg

fbshipit-source-id: 1fdd0de7f77744a2f2c9c063d6a56d499e8ac7b3