[examples] Fix resnet-verify
Summary:
This commit fixes the resnet-verify example program. The code that inserts
a SaveNode after every intermediate node in a network in order to compare
its output across backends did not handle nodes that have more than one output
(e.g. MaxPool) and this caused the program to crash in Debug mode. In release mode,
the program silently verifies only the first result. This commit fixes that by
modifying the HookedFunction struct so that it contains a *list* of SaveNodes and
their corresponding Placeholders, one for each result of the hooked node.

Testing:
1) Ran the binary in debug mode - it doesn't crash anymore.
2) Ran the binary in release mode - it still runs to completion.