Add numeric equivalence check to GraphOptz.optimizeBatchN… (#3673)
Summary:
…ormAfterConvButVarReused

Also simplify test and make it more similar to others in the file.
Add numerical equivalence check to GraphOptz.optimizeBatchNormAfterConvButVarReused to ensure that the optimized graph gives the same numerical results as the unoptimized graph.
Pull Request resolved: https://github.com/pytorch/glow/pull/3673

Test Plan:
Build and run tests a few times to make sure it's not flaky

```
$ for _ in {0..10}; do ninja test; done
```

Differential Revision: D18104168

Pulled By: mgraczyk

fbshipit-source-id: 8af51f12e266c9d8afc53d806ae2c00f5843bf1f