Add AdaptiveAvgPool on CPU backend (#3734)
Summary:
This adds only fp32 support on the CPU because quantized support will require downscaling from int32 to int8 and the scale used to compute this needs to be an input to the jit kernel but it also needs to be divided by the height * width of the area being averaged over which is only known inside of the jit kernel. This could be solved by precomputing all of the scales for the output ahead of time and passing them into the jit kernel as a constant array of size height*width of the output with `emitConstSizeTArray` but this doesn't seem very pleasant.

Documentation:
doxygen
Pull Request resolved: https://github.com/pytorch/glow/pull/3734

Test Plan:
enable CPUOperatorTests
try it on pytorch resnet_example.py

Differential Revision: D18329389

Pulled By: jackm321

fbshipit-source-id: a76bd6a6ae5f1d3ec857f4f2a935e55349e8e45d