[Operator] Add SparseLengthSumGradNode and Inst
Summary:
This commit adds SparseLengthsSumGradNode and Inst. SparseLengthsSum is currently
lowered to SparseLengthsWeightedSum with all weights set to 1, but some backends
might want to opt out of this lowering. Those backends will then need
SparseLengthsSumGradNode and the corresponding Inst.

Test Plan:
Not much to test in this commit but all unit tests pass.