Quantized FC node (#3737)
Summary:
Testing to see if matmul -> ba (difference in order of bias addition and requantization) can account for the small differences from pytorch quantized linear. It seem that when bias and scales are small, this is in fact much more accurate but is more subject to rounding differences when scales are large.

Documentation:
doxygen
Pull Request resolved: https://github.com/pytorch/glow/pull/3737

Test Plan: unit tests

Differential Revision: D18421538

Pulled By: jackm321

fbshipit-source-id: cb5d87cf6ad21031642a2b9ba2436a1878037425