[GraphOptz] Add helper to check whether a Node is in a Function
Summary:
The tests in GraphOptzTest use the helper predicate IsSameNodeAddress in
conjunction with std::find_if to check whether a Node is still in a
Function after it has been optimized. This commit introduces a helper
function functionContainsNode that contains this logic and replaces all of
the calls to std::find_if with IsSameNodeAddress with this new function
for improved readability. This function will also be used frequently
while adding numerical equivalence checks to more tests.

Test Plan:
All unit tests pass.