Fix broken DeviceManagerTest test cases (#3820)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/3820

This commit fixes some broken DeviceManagerTest test cases. In the case of the
ReuseModule and MultiModule tests, the tests did not call `ensureOnHost` before
checking the contents of output Tensors. In the case of the
CanHandleDeviceResidentTensors test, the runFunction implementation of the
DeviceManager subclass used in the test is a no-op, so the output Tensor of the
Function used for the test is not set at all. Thus, there is no reason to
expect it to be equal to a Tensor produced by applying a host-side
implementation of the Function to the same input.

Reviewed By: nickgg

Differential Revision: D18689157

fbshipit-source-id: 1764250ecdbd84a743c4fe410e7dea2d0ed04eb2